Add FriendRequest and Notifications
Create base for requests and notification models.

Notifications have a single inheritance table with polymorphic
associations. All notifications can be loaded for a user with a single
query, but each subclass of notification can specify its own display
partial in its partial method. The corresponding notifiable model can be
any type of request (friend request, invite, etc.) or notice (comment on
post, birthday, poke, etc.).

Friend requests contain foreign keys linking the sender and receiver, as
well as the status of the request (unanswered, accepted or declined).
Each request validates that the sender and receiver are not the same
user on creation.