Add Like Model
The like model belongs to two other models: a user and a polymorphic
likeable model.

In order to make the LikesController flexible, each create action of
likes is nested in the likeable resource, with an additional parameter
of likeable_type sent with the route. The LikesController uses this get
the likeable model and build the like from there.

The delete action only requires the like model's id.