Add ProfilesController
After a user creates a new account, they can then create their
public-facing profile.

In order to prevent users without profiles from interacting with the
site (since they would have no public-facing name), a before_action in
ApplicationController checks for the existence of their profile, and if
the profile does not exist, send them to profile#new.

Plan to optimize this in the future.