Fix error in friend request display on User show
A variable declaration for an existing friend_request was missing on the
User Show page causing an error.

The if statement searches for an existing friend_request using find_by,
and if the request is found, stores it in the friend_request variable,
which is used by the below code to create forms for updating the friend
request.