Fix Invalid Authenticity Token Error
A strange error occured only after navigating to the Notifications page
and trying to archive or mark a notification as read. After reloading
the page and trying again, the token somehow worked.

The error was fixed by updating the path in the form_with method from
"notifications/#{notification.id}" to "/notifications/#{notification.id}".

Why did that cause an authenticity token error in the first place? Why
did it work on a subsequent reload of the page? It's a mystery. But at
least it's over now.