Update handling of failed user creation in Omniauth
The user is sent to a new registration page if a new user isn't created
after Omniauth login attempt, but the flash doesn't tell the user why.

Two possible handling situations: a user with email already exists or
something went wrong. For the first, inform the user that an account
exists via the flash and send to sign in page. For the second, inform
the user that something went wrong and send them to new registration
page.