Add profile_picture_url helper
profile_picture_url helper takes a user parameter and a dimensions array
parameter and returns the url to the user's profile photo in the
requested size if it exists.

If there is no profile picture or the attached avatar is not an image,
it returns the url to a placeholder image.

The helper has replaced several instances of redundant logic in
user-related views and partials.